\name{check.ci.cl-internal}
\alias{check.ci.cl}
\title{Check confidence level for AUC's confidence intervals}
\description{Checks the validity of the value entered as confidence level for computing the confidence intervals for the AUC.}
\usage{check.ci.cl(ci.cl)}
\arguments{
  \item{ci.cl}{confidence level at which the confidence intervals for the AUC will be computed.}
}
\details{
Verifies that the value entered as confidence level ranges between 0 and 1. The 0.95 confidence level is taken as default.
}
\value{A list with two components:
\item{ci.cl}{value entered as confidence level for the AUC.}
\item{message}{table with the warning messages generated by the function.}
}
