\name{check.grid-internal}
\alias{check.grid}
\title{Check grid}
\description{Cheking of the parameter grid.}
\usage{check.grid(grid)}
\arguments{
  \item{grid}{ grid size for computing the ROC curve estimate. The default value is 1000.}
}
\details{
Verifies if the parameter entered as grid is a numerical value greater than 0.
}
\value{A list with two components:
\item{grid}{ grid size entered.}
\item{message}{table with the warning messages generated by the function.}
}

