\name{check.nboots-internal}
\alias{check.nboots}
\title{Check number of bootstrap samples}
\description{This function checks if the value entered as number of bootstrap samples is correct.}
\usage{check.nboots(nboots)}
\arguments{
  \item{nboots}{ number of bootstrap samples to be run. The default value is 500.}
}
\value{A list with two components:
\item{nboots}{ value entered as number of bootstrap samples.}
\item{message}{table with the warning messages generated by the function.}
}
