\name{check.ncpus-internal}
\alias{check.ncpus}
\title{Check number of CPUs}
\description{Checks the number of CPUs to be used when parallel computing is performed. The default value is 1 and the maximum is 2.}
\usage{check.ncpus(ncpus)}
\arguments{
  \item{ncpus}{ number of CPUs to be used when performing parallel computing.}
}
\value{A list with two components:
\item{ncpus}{ value entered as number of CPUs chosen.}
\item{message}{table with the warning messages generated by the function.}
}
