\name{check.time-internal}
\alias{check.time}
\title{Check time}
\description{Cheking of the parameter time.}
\usage{check.time(time)}
\arguments{
  \item{time}{ point of time at which the ROC curve estimate in prognosis scenarios will be computed. It is mandatory in this scenario.}
}
\value{A list with two components:
\item{time}{ value entered as time.}
\item{message}{table with the warning messages generated by the function.}
}

