\name{check_tim-internal}
\alias{check_tim}
\title{Check tim}
\description{Cheking of the parameter time.}
\usage{check_tim(time)}
\arguments{
  \item{time}{ point of time at which the ROC curve estimate in prognosis scenarios will be computed. It is mandatory in this scenario.}
}
\value{A list with two components:
\item{time}{ value entered as time.}
\item{message}{table with the warning messages generated by the function.}
}

